refactor BasePage (delete throw Exceptions), add new tests, add Login annotation, create new xml file, create data generator, add Header class, add wrappers
