+++ src/test/java/tests/LoginPageTest.java
@@ -1,13 +1,16 @@
 package tests;
 
 import lombok.extern.log4j.Log4j2;
-import org.testng.Assert;
 import org.testng.annotations.BeforeMethod;
 import org.testng.annotations.Test;
 import pages.LoginPage;
 import pages.ProjectsPage;
 import tests.base.BaseTest;
+import tests.base.annotations.Login;
+import utils.DataGenerator;
 import utils.constants.Constants;
+
+import static org.testng.Assert.assertTrue;
 
 @Log4j2
 public class LoginPageTest extends BaseTest {
@@ -18,17 +21,54 @@ public class LoginPageTest extends BaseTest {
     }
 
     @Test
+    @Login(doLogIn = false)
     public void testLoginWithAValidUsernameAndPassword() {
-        log.info("Precondition. Navigate to login page as a non-logged user");
         LoginPage loginPage = new LoginPage(driver);
-
-        log.info("Step 1. Fill the User Field and the Password field with valid data then click on 'Login' button");
         loginPage
                 .fillUserName(Constants.USERNAME)
                 .fillPassword(Constants.PASSWORD)
                 .clickLoginButton();
-        log.info("Attempt to confirm User logs in and is directed to the Projects Page");
         ProjectsPage projectsPage = new ProjectsPage(driver);
-        Assert.assertTrue(projectsPage.isPageLoaded(), "Project Page was not loaded");
+        assertTrue(projectsPage.isPageLoaded(), "Project Page was not loaded");
+    }
+
+    @Test
+    @Login(doLogIn = false)
+    public void testLoginWithAValidUsernameAndInvalidPassword() {
+        String incorrectPassword = DataGenerator.generateRandomAlphaNumericString(10);
+
+        LoginPage loginPage = new LoginPage(driver);
+        loginPage
+                .fillUserName(Constants.USERNAME)
+                .fillPassword(incorrectPassword)
+                .clickLoginButton();
+        assertTrue(loginPage.isInputErrorVisible(), "Message about wrong entered credentials isn't shown");
+    }
+
+    @Test
+    @Login(doLogIn = false)
+    public void testLoginWithInvalidUsernameAndValidPassword() {
+        String incorrectUsername = DataGenerator.generateRandomAlphaNumericString(6) + "@gmail.com";
+
+        LoginPage loginPage = new LoginPage(driver);
+        loginPage
+                .fillUserName(incorrectUsername)
+                .fillPassword(Constants.PASSWORD)
+                .clickLoginButton();
+        assertTrue(loginPage.isInputErrorVisible(), "Message about wrong entered credentials isn't shown");
+    }
+
+    @Test
+    @Login(doLogIn = false)
+    public void testLoginWithInvalidUsernameAndInvalidPassword() {
+        String incorrectPassword = DataGenerator.generateRandomAlphaNumericString(10);
+        String incorrectUsername = DataGenerator.generateRandomAlphaNumericString(6) + "@gmail.com";
+
+        LoginPage loginPage = new LoginPage(driver);
+        loginPage
+                .fillUserName(incorrectUsername)
+                .fillPassword(incorrectPassword)
+                .clickLoginButton();
+        assertTrue(loginPage.isInputErrorVisible(), "Message about wrong entered credentials isn't shown");
     }
 }
