+++ src/test/java/tests/base/annotations/Login.java
@@ -1,0 +1,16 @@
+package tests.base.annotations;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+@Retention(RetentionPolicy.RUNTIME)
+@Target(ElementType.METHOD)
+public @interface Login {
+    boolean doLogIn() default true;
+
+    String login() default "";
+
+    String password() default "";
+}
