+++ src/test/java/utils/DataGenerator.java
@@ -1,0 +1,25 @@
+package utils;
+
+import org.apache.commons.lang3.RandomStringUtils;
+
+import java.util.Random;
+
+public class DataGenerator {
+
+    public static String generateRandomAlphaNumericString(int length) {
+        return RandomStringUtils.random(length, true, true);
+    }
+
+    public static String generateRandomNumericString(int length) {
+        return RandomStringUtils.random(length, false, true);
+    }
+
+    public static String generateRandomAlphabeticString(int length) {
+        return RandomStringUtils.random(length, true, false);
+    }
+
+    public static boolean generateRandomBoolean() {
+        Random random = new Random();
+        return random.nextBoolean();
+    }
+}
