--- src/test/resources/MainSuite.xml
@@ -1,25 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!DOCTYPE suite SYSTEM "http://testng.org/testng-1.0.dtd" >
-<suite name="Global suite">
-
-    <test name="#1 dont work git actions :( ">
-        <classes>
-            <class name="tests.ProjectTest">
-                <methods>
-                    <include name="createProjectTest"/>
-                </methods>
-            </class>
-        </classes>
-    </test>
-
-    <test name="#2 dont work git actions. Its sad :( ">
-        <classes>
-            <class name="tests.LoginPageTest">
-                <methods>
-                    <include name="testLoginWithAValidUsernameAndPassword"/>
-                </methods>
-            </class>
-        </classes>
-    </test>
-
-</suite>
