+++ src/test/resources/testXml/GlobalUISuite.xml
@@ -1,0 +1,10 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!DOCTYPE suite SYSTEM "http://testng.org/testng-1.0.dtd" >
+<suite name="Global UI suite">
+
+    <suite-files>
+        <suite-file path="web/LoginPage.xml"/>
+        <suite-file path="web/ProjectsPage.xml"/>
+    </suite-files>
+
+</suite>
